+++ app/lib/ui/components/form_input/phone_number/phone_number_input_demo_screen.dart
@@ -244,12 +244,14 @@ class _PhoneNumberInputDemoState extends State<_PhoneNumberInputDemo> {
               : null,
           hasPrefix: customizationState.hasPrefix,
           prefixIcon: customizationState.hasLeadingIcon
-              ? AppAssets.icons.deviceSmartphone(themeController)
+              ? OudsTextInputPrefixIcon(
+                  icon: AppAssets.icons.deviceSmartphone(themeController),
+                )
               : null,
           errorText: customizationState.hasError
               ? context.l10n.app_components_phoneNumberInput_error_label
               : null,
-          loader: customizationState.hasLoader,
+          loader: customizationState.loader,
           outlined: customizationState.hasOutlined,
           constrainedMaxWidth: customizationState.hasConstrainedMaxWidth
               ? true